Move cr2_to_pgm to io
