+++ pocs/utils/images/conversions.py
@@ -5,8 +5,6 @@ from astropy.io import fits
 from pprint import pprint
 from warnings import warn
 
-from dateutil import parser as date_parser
-
 from pocs.utils import current_time
 from pocs.utils import error
 from pocs.version import version
@@ -95,141 +93,6 @@ def process_cr2(cr2_fname, to_fits=True, fits_headers={}, solve=False, make_pret
     Returns:
         fits.PrimaryHDU:   FITS file
     """
-
-
-def cr2_to_fits(cr2_fname, fits_fname=None, clobber=False, fits_headers={}, remove_cr2=False, **kwargs):
-    """ Convert a CR2 file to FITS
-
-    This is a convenience function that first converts the CR2 to PGM via `cr2_to_pgm`. Also adds keyword headers
-    to the FITS file.
-
-    Note:
-        The intermediate PGM file is automatically removed
-
-    Arguments:
-        cr2_fname {str} -- Name of CR2 file to be converted
-        **kwargs {dict} -- Additional keywords to be used
-
-    Keyword Arguments:
-        fits_fname {str} -- Name of FITS file to output. If None (default), the `cr2_fname` is used
-            as base (default: {None})
-        clobber {bool} -- A bool indicating if existing FITS should be clobbered (default: {False})
-        fits_headers {dict} -- Header values to be saved with the FITS, by default includes the EXIF
-            info from the CR2 (default: {{}})
-        remove_cr2 {bool} -- A bool indicating if the CR2 should be removed (default: {False})
-
-    """
-
-    verbose = kwargs.get('verbose', False)
-
-    if fits_fname is None:
-        fits_fname = cr2_fname.replace('.cr2', '.fits')
-
-    if not os.path.exists(fits_fname) or clobber:
-        if verbose:
-            print("Converting CR2 to PGM: {}".format(cr2_fname))
-
-        # Convert the CR2 to a PGM file then delete PGM
-        pgm = read_pgm(cr2_to_pgm(cr2_fname), remove_after=True)
-
-        # Add the EXIF information from the CR2 file
-        exif = read_exif(cr2_fname)
-
-        # Set the PGM as the primary data for the FITS file
-        hdu = fits.PrimaryHDU(pgm)
-
-        # Set some default headers
-        hdu.header.set('FILTER', 'RGGB')
-        hdu.header.set('ISO', exif.get('ISO', ''))
-        hdu.header.set('EXPTIME', exif.get('ExposureTime', ''))
-        hdu.header.set('CAMTEMP', exif.get('CameraTemperature', ''))
-        hdu.header.set('CIRCCONF', exif.get('CircleOfConfusion', ''))
-        hdu.header.set('COLORTMP', exif.get('ColorTempMeasured', ''))
-        hdu.header.set('DATE-OBS', date_parser.parse(exif.get('DateTimeOriginal', '')).isoformat())
-        hdu.header.set('FILENAME', exif.get('FileName', ''))
-        hdu.header.set('INTSN', exif.get('InternalSerialNumber', ''))
-        hdu.header.set('CAMSN', exif.get('SerialNumber', ''))
-        hdu.header.set('MEASEV', exif.get('MeasuredEV', ''))
-        hdu.header.set('MEASEV2', exif.get('MeasuredEV2', ''))
-        hdu.header.set('MEASRGGB', exif.get('MeasuredRGGB', ''))
-        hdu.header.set('WHTLVLN', exif.get('NormalWhiteLevel', ''))
-        hdu.header.set('WHTLVLS', exif.get('SpecularWhiteLevel', ''))
-        hdu.header.set('REDBAL', exif.get('RedBalance', ''))
-        hdu.header.set('BLUEBAL', exif.get('BlueBalance', ''))
-        hdu.header.set('WBRGGB', exif.get('WB_RGGBLevelAsShot', ''))
-
-        if verbose:
-            print("Adding provided FITS header")
-
-        for key, value in fits_headers.items():
-            try:
-                hdu.header.set(key.upper()[0: 8], "{}".format(value))
-            except:
-                pass
-
-        try:
-            if verbose:
-                print("Saving fits file to: {}".format(fits_fname))
-
-            hdu.writeto(fits_fname, output_verify='silentfix', clobber=clobber)
-        except Exception as e:
-            warn("Problem writing FITS file: {}".format(e))
-        else:
-            if remove_cr2:
-                os.unlink(cr2_fname)
-
-    return fits_fname
-
-
-def cr2_to_pgm(cr2_fname, pgm_fname=None, dcraw='/usr/bin/dcraw', clobber=True, **kwargs):
-    """ Convert CR2 file to PGM
-
-    Converts a raw Canon CR2 file to a netpbm PGM file via `dcraw`. Assumes `dcraw` is installed on the system
-
-    Note:
-        This is a blocking call
-
-    Arguments:
-        cr2_fname {str} -- Name of CR2 file to convert
-        **kwargs {dict} -- Additional keywords to pass to script
-
-    Keyword Arguments:
-        pgm_fname {str} -- Name of PGM file to output, if None (default) then use same name as CR2 (default: {None})
-        dcraw {str} -- Path to installed `dcraw` (default: {'/usr/bin/dcraw'})
-        clobber {bool} -- A bool indicating if existing PGM should be clobbered (default: {True})
-
-    Returns:
-        str -- Filename of PGM that was created
-
-    """
-    assert os.path.exists(dcraw), "dcraw does not exist at location {}".format(dcraw)
-    assert os.path.exists(cr2_fname), "cr2 file does not exist at location {}".format(cr2_fname)
-
-    verbose = kwargs.get('verbose', False)
-
-    if pgm_fname is None:
-        pgm_fname = cr2_fname.replace('.cr2', '.pgm')
-
-    if os.path.exists(pgm_fname) and not clobber:
-        if verbose:
-            print("PGM file exists and clobber=False, returning existing file: {}".format(pgm_fname))
-    else:
-        try:
-            # Build the command for this file
-            command = '{} -t 0 -D -4 {}'.format(dcraw, cr2_fname)
-            cmd_list = command.split()
-            if verbose:
-                print("PGM Conversion command: \n {}".format(cmd_list))
-
-            # Run the command
-            if subprocess.check_call(cmd_list) == 0:
-                if verbose:
-                    print("PGM Conversion command successful")
-
-        except subprocess.CalledProcessError as err:
-            raise InvalidSystemCommand(msg="File: {} \n err: {}".format(cr2_fname, err))
-
-    return pgm_fname
 
 
 def make_pretty_image(fname, timeout=15, **kwargs):
